this lays the groundwork for more authenticated functions
- extracted all the login functions from views/sis
- removed views/sis/login
- added login capabilities to settings
