+++ lib/financials.js
@@ -1,0 +1,10 @@
+// async function weeklyMealsRemaining() {
+//   let {username, password} = await loadLoginCredentials()
+//   let form = buildFormData({
+//     username: username,
+//     password: password,
+//   })
+//   let page = await api.post('/apps/olecard/checkbalance/authenticate.cfm', {body: form})
+//   console.log(page)
+//   return page
+// }
